Move tunnel configuration generation into TunnelSettingsV2 extension
